Bugfix 6732: Fix incorrect handling of single-value leaf nodes.
It turns out that a script needs to call the returnValue function,
not the returnValues function in Vyatta::Config to get the value
of a single-valued node.  The returnValues function is reserved
for the exclusive use of multi-valued nodes, and will just return
an empty value array if called on a single-valued leaf node.