Bugfix 6732: Fixed more API issues.
A perl script needs to use different API functions to get the value of
a single-valued node from within config mode than it needs to use
outside of config mode.  And those two API functions are different
from the two the script has to use to get the value or values of a
multi-valued node, which are also different depending on whether the
caller is running inside or outside of config mode.
(cherry picked from commit eef99a8264586ff308e6adf224847baf7fd84029)